Changed the selection of affiliation section
The previous implementation was always selecting the second info
section from the mandate page. However, this is not always the case
because many deputies have other info sections on the page before the
affiliations section.

The current logic iterates over all sections and looks for
Resources.PoliticalParty or Resources.PoliticalGroup strings in the
section title.